notifications work just need them happening between users. Also fixed some file structure things. Moved unimplemented classes into their own folder.
