+++ app/src/main/java/com/example/schat/Chat/ChatListAdapter.java
@@ -10,10 +10,7 @@ import android.view.ViewGroup;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
-import com.example.schat.ChatActivity;
 import com.example.schat.R;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.FirebaseDatabase;
 
 import java.util.ArrayList;
