+++ app/src/main/java/com/example/schat/MainFeedActivity.java
@@ -15,6 +15,10 @@ import android.widget.LinearLayout;
 import com.example.schat.Chat.ChatObject;
 import com.example.schat.User.FindUserActivity;
 import com.example.schat.Chat.ChatListAdapter;
+import com.example.schat.Utils.SendNotification;
+import com.example.schat.unimplemented.ContactsActivity;
+import com.example.schat.unimplemented.ProfileActivity;
+import com.example.schat.unimplemented.SettingsActivity;
 import com.facebook.drawee.backends.pipeline.Fresco;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DataSnapshot;
@@ -22,6 +26,7 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+import com.onesignal.OneSignal;
 
 import java.util.ArrayList;
 
@@ -37,6 +42,18 @@ public class MainFeedActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main_feed);
+        OneSignal.startInit(this).init();
+        OneSignal.setSubscription(true);
+        OneSignal.idsAvailable(new OneSignal.IdsAvailableHandler() {
+            @Override
+            public void idsAvailable(String userId, String registrationId) {
+                FirebaseDatabase.getInstance().getReference().child("user").child(FirebaseAuth.getInstance().getUid()).child("notificationKey").setValue(userId);
+            }
+        });
+        OneSignal.setInFocusDisplaying(OneSignal.OSInFocusDisplayOption.Notification);
+
+        new SendNotification("message 1", "heading 1", null);
+
         Fresco.initialize(this);
 
         // contact search
@@ -54,6 +71,7 @@ public class MainFeedActivity extends AppCompatActivity {
         mLogout.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                OneSignal.setSubscription(false);
                 FirebaseAuth.getInstance().signOut();
                 Intent intent = new Intent(getApplicationContext(), WelcomeScreen.class);
                 intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
