+++ app/src/main/java/com/example/schat/unimplemented/ProfileActivity.java
@@ -1,13 +1,15 @@
-package com.example.schat;
+package com.example.schat.unimplemented;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class SettingsActivity extends AppCompatActivity {
+import com.example.schat.R;
+
+public class ProfileActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings);
+        setContentView(R.layout.activity_profile);
     }
 }
