+++ app/src/main/java/com/example/schat/unimplemented/SettingsActivity.java
@@ -1,13 +1,15 @@
-package com.example.schat;
+package com.example.schat.unimplemented;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class ProfileActivity extends AppCompatActivity {
+import com.example.schat.R;
+
+public class SettingsActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile);
+        setContentView(R.layout.activity_settings);
     }
 }
