+++ app/src/main/res/layout/activity_chat.xml
@@ -3,7 +3,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".ChatActivity">
+tools:context=".Chat.ChatActivity">
 
 <android.support.v7.widget.RecyclerView
     android:layout_width="match_parent"
